code splitting and refactoring
