+++ stack-cli/src/main.rs
@@ -1,11 +1,8 @@
 use core::fmt;
 use std::{
-  collections::HashMap,
   io::{self, prelude::Write, Read},
-  mem,
   path::{Path, PathBuf},
-  rc::Rc,
-  sync::{Arc, Mutex},
+  sync::Arc,
 };
 
 use clap::Parser;
@@ -26,8 +23,8 @@ use notify::{
   Config, Event, EventKind, RecommendedWatcher, RecursiveMode, Watcher,
 };
 use reedline::{DefaultPrompt, DefaultPromptSegment, Reedline, Signal};
+use stack_cli::server::listen;
 use stack_core::prelude::*;
-use ws::Message;
 
 fn main() {
   let cli = Cli::parse();
@@ -227,202 +224,7 @@ fn main() {
         }
       }
     }
-    Subcommand::Serve => {
-      use serde::{Deserialize, Serialize};
-      use ws::listen;
-
-      let eng_mutex = Rc::new(Mutex::new(Engine::new()));
-      let ctx_mutex = Rc::new(Mutex::new(Context::new()));
-
-      #[derive(Debug, Clone, Deserialize)]
-      #[serde(tag = "type", content = "code", rename_all = "lowercase")]
-      enum Incoming {
-        /// Run code within the existing engine
-        Run(String),
-        /// Run code within a new engine
-        #[serde(rename = "run_new")]
-        RunNew(String),
-
-        /// Exports the stack
-        Stack,
-        /// Exports the current scope
-        Scope,
-
-        /// Clear the stack
-        #[serde(rename = "clear_stack")]
-        ClearStack,
-        /// Clear the scope
-        #[serde(rename = "clear_scope")]
-        ClearScope,
-        /// Clear everything
-        #[serde(rename = "clear")]
-        ClearAll,
-      }
-
-      #[derive(Debug, Clone, Serialize)]
-      #[serde(tag = "error", content = "value", rename_all = "lowercase")]
-      enum OutgoingError {
-        /// Error from the Engine
-        #[serde(rename = "run_error")]
-        RunError(RunError),
-
-        /// Error from the command reader
-        #[serde(rename = "command_error")]
-        CommandError(String),
-      }
-
-      #[derive(Debug, Clone, Serialize)]
-      #[serde(tag = "status", content = "value", rename_all = "lowercase")]
-      enum Outgoing {
-        /// A Expr
-        #[serde(rename = "ok")]
-        Single(Expr),
-
-        /// A Null Response
-        #[serde(rename = "ok")]
-        Null(()),
-
-        /// A Vec of Exprs
-        #[serde(rename = "ok")]
-        Many(Vec<Expr>),
-
-        /// A Map of Exprs
-        #[serde(rename = "ok")]
-        Map(HashMap<String, Expr>),
-
-        /// An Error
-        Error(OutgoingError),
-      }
-
-      listen("127.0.0.1:5001", |out| {
-        let eng_mutex = eng_mutex.clone();
-        let ctx_mutex = ctx_mutex.clone();
-
-        move |msg| {
-          if let Message::Text(string) = msg {
-            let request = serde_json::from_str::<Incoming>(&string);
-
-            match request {
-              Ok(incoming) => match incoming {
-                Incoming::RunNew(code) => {
-                  let source = Source::new("runner", code);
-                  let mut lexer = Lexer::new(source);
-                  let exprs = parse(&mut lexer).unwrap();
-
-                  match (eng_mutex.try_lock(), ctx_mutex.try_lock()) {
-                    (Ok(engine), Ok(mut guard)) => {
-                      let _ = mem::replace(&mut *guard, Context::new());
-
-                      let context = mem::take(&mut *guard);
-                      let result = engine.run(context, exprs);
-
-                      match result {
-                        Ok(ctx) => {
-                          *guard = ctx;
-
-                          match guard.stack().last().cloned() {
-                            Some(expr) => out.send(
-                              serde_json::to_string(&Outgoing::Single(expr))
-                                .unwrap(),
-                            ),
-                            None => out.send(
-                              serde_json::to_string(&Outgoing::Null(()))
-                                .unwrap(),
-                            ),
-                          }
-                        }
-                        Err(error) => out.send(
-                          serde_json::to_string(&Outgoing::Error(
-                            OutgoingError::RunError(error),
-                          ))
-                          .unwrap(),
-                        ),
-                      }
-                    }
-                    _ => todo!("mutex not lock"),
-                  }
-                }
-                Incoming::Run(code) => {
-                  let source = Source::new("runner", code);
-                  let mut lexer = Lexer::new(source);
-                  let exprs = parse(&mut lexer).unwrap();
-
-                  match (eng_mutex.try_lock(), ctx_mutex.try_lock()) {
-                    (Ok(engine), Ok(mut guard)) => {
-                      let context = mem::take(&mut *guard);
-                      let result = engine.run(context, exprs);
-
-                      match result {
-                        Ok(ctx) => {
-                          *guard = ctx;
-
-                          let expr = guard
-                            .stack()
-                            .last()
-                            .cloned()
-                            .unwrap_or_else(|| ExprKind::Nil.into());
-
-                          out.send(
-                            serde_json::to_string(&Outgoing::Single(expr))
-                              .unwrap(),
-                          )
-                        }
-                        Err(error) => out.send(
-                          serde_json::to_string(&Outgoing::Error(
-                            OutgoingError::RunError(error),
-                          ))
-                          .unwrap(),
-                        ),
-                      }
-                    }
-                    _ => todo!("mutex not lock"),
-                  }
-                }
-
-                Incoming::Stack => match ctx_mutex.try_lock() {
-                  Ok(context) => out.send(
-                    serde_json::to_string(&Outgoing::Many(
-                      context.stack().to_vec(),
-                    ))
-                    .unwrap(),
-                  ),
-                  Err(_) => todo!(),
-                },
-                Incoming::Scope => match ctx_mutex.try_lock() {
-                  Ok(context) => {
-                    let mut scope: HashMap<String, Expr> = HashMap::new();
-                    for (k, v) in context.scope().items.iter() {
-                      scope.insert(
-                        k.to_string(),
-                        v.borrow().val().clone().unwrap(),
-                      );
-                    }
-
-                    out.send(
-                      serde_json::to_string(&Outgoing::Map(scope)).unwrap(),
-                    )
-                  }
-                  Err(_) => todo!(),
-                },
-
-                Incoming::ClearStack => todo!(),
-                Incoming::ClearScope => todo!(),
-                Incoming::ClearAll => todo!(),
-              },
-              Err(parse_error) => out.send(
-                serde_json::to_string(&Outgoing::Error(
-                  OutgoingError::CommandError(parse_error.to_string()),
-                ))
-                .unwrap(),
-              ),
-            }
-          } else {
-            todo!("message not text")
-          }
-        }
-      })
-      .unwrap();
-    }
+    Subcommand::Serve => listen(),
   }
 }
